We need also to hide the edit and delete buttons even for the other users when they are in posts which not belong them.
If we manually go to the URL of edit we still can edit so we need to prevent that by going to the edit in the PostController by adding if statement to boda edit and destroy for checking the correct user.

+++ app/Http/Controllers/PostsController.php
@@ -89,6 +89,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view ('posts.edit')->with('post',$post);
     }
 
@@ -124,8 +130,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
 
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        $post->delete();
         return redirect ('/posts')->with('success','Post Removed');
     }
 }
